# tpl: Modify 'Querify' to take lone slice argument