@moorereason 2021-03-07
I don't think this will do what you intend.  To resolve the issue, `len(params)` will equal `1` since `params` will be something like `[]interface{ []string{"key","value"} }`.  I would suggest adding a test for the scenario you want to pass--we're almost always require tests anyway before we'll merge.

My hunch is that you could do something like this instead:

```
var (
  vals map[string]interface{}
  err error
)

// check for special case of a single slice param
if len(params) == 1 && reflect.TypeOf(params[0]).Kind() == reflect.Slice {
    vals, err = ns.Dictionary(params[0]...)
} else {
    vals, err = ns.Dictionary(params...)
}
if err != nil {
    // ...
```

Reflection is one of the trickier parts of the Go stdlib, so I may be a little off with the above...but it's probably close.  😄