@importhuman 2021-03-07
So if I understand correctly, what I've done so far only solves the issue of key/value pairs being unordered due to map, but I haven't implemented the lone slice argument as was the issue.

Should I add test cases to the original test only, considering it's the same function?