@moorereason 2021-03-07
> Should I add test cases to the original test only, considering it's the same function?

Yes