@moorereason 2021-03-10
Are you sure there's a problem with key ordering in the current implementation?  If there were, wouldn't our existing tests be failing randomly with indeterminate results?  I'm wondering if you're solving a problem that doesn't exist.