@importhuman 2021-03-11
@moorereason Yes I was wondering the same thing about the tests, but as mentioned [here](https://github.com/gohugoio/hugo/issues/6735#issuecomment-572704093) it couldn't be a Go map, so I stayed away from maps and implemented a different solution.