@moorereason 2021-03-13
@importhuman,
I've modified your PR locally with some optimizations, but I'll let you try to beat me if you're up for it.  :smile:  Otherwise, I'll just show you my work (but you'll learn more if you try it yourself).  No shame if you don't have time for this.

I pushed a couple benchmarks to your branch.  To run them, I'd use the following to get a baseline of your PR:

```
$ go test -bench='BenchmarkQuerify.*' -count=5 ./tpl/collections | tee bench.querify
WARNING: calling IsSet with unsupported type "struct" (time.Time) will always return false.

WARNING: calling IsSet with unsupported type "invalid" (<nil>) will always return false.

goos: linux
goarch: amd64
pkg: github.com/gohugoio/hugo/tpl/collections
cpu: Intel(R) Core(TM) i7-7600U CPU @ 2.80GHz
BenchmarkQuerify-2                643951              1975 ns/op             464 B/op         18 allocs/op
BenchmarkQuerify-2                689896              1837 ns/op             464 B/op         18 allocs/op
BenchmarkQuerify-2                696940              1833 ns/op             464 B/op         18 allocs/op
BenchmarkQuerify-2                748134              2201 ns/op             464 B/op         18 allocs/op
BenchmarkQuerify-2                738346              1846 ns/op             464 B/op         18 allocs/op
BenchmarkQuerifySlice-2           454816              2263 ns/op             552 B/op         23 allocs/op
BenchmarkQuerifySlice-2           473956              2322 ns/op             552 B/op         23 allocs/op
BenchmarkQuerifySlice-2           615118              2245 ns/op             552 B/op         23 allocs/op
BenchmarkQuerifySlice-2           568017              2278 ns/op             552 B/op         23 allocs/op
BenchmarkQuerifySlice-2           441758              2330 ns/op             552 B/op         23 allocs/op
PASS
ok      github.com/gohugoio/hugo/tpl/collections        13.997s
```
Run it again after making changes, write the output to another file, and then use [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to compare results.  Here's your version compared to mine:

```
$ benchstat bench.querify worktrees/iss6735-querify/bench.querify
name            old time/op    new time/op    delta
Querify-2         1.94µs ±14%    1.02µs ±17%  -47.55%  (p=0.008 n=5+5)
QuerifySlice-2    2.29µs ± 2%    0.91µs ±17%  -60.40%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
Querify-2           464B ± 0%      152B ± 0%  -67.24%  (p=0.008 n=5+5)
QuerifySlice-2      552B ± 0%      176B ± 0%  -68.12%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
Querify-2           18.0 ± 0%       9.0 ± 0%  -50.00%  (p=0.008 n=5+5)
QuerifySlice-2      23.0 ± 0%       9.0 ± 0%  -60.87%  (p=0.008 n=5+5)
```
Focus your efforts on reducing memory allocations.  The time results may not be super-accurate since I'm running these on a VM on my laptop that will mess with the CPU clock speed as it sees fit (but they're useful nonetheless).