@importhuman 2021-03-13
@moorereason I'd like to try it myself, but I didn't understand this step:
> Run it again after making changes, write the output to another file
What exactly am I writing to another file, and where (as in the branch, etcetera). Also, I haven't read the Go spec yet, should I refer to that or some other resource for help on what will help memory allocations? I'm guessing I'll also have to study up a bit on the O(n) notation stuff?