@moorereason 2021-03-13
Man, my language is imprecise. I mean tee the output to another file on the command line. benchstat just needs two files with benchmark results to compare. 

O(n)? No, it's not that complicated really. Hint: adding items to pairedSlice requires an allocation. Is it necessary?