@importhuman 2021-03-14
@moorereason This is all I've managed:

```
~/s/hugo ❯❯❯ benchstat bench.querify bench2.querify
name            old time/op    new time/op    delta
Querify-4         3.25µs ± 3%    2.94µs ± 3%   -9.47%  (p=0.008 n=5+5)
QuerifySlice-4    3.96µs ± 2%    3.75µs ± 1%   -5.33%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
Querify-4           464B ± 0%      440B ± 0%   -5.17%  (p=0.008 n=5+5)
QuerifySlice-4      552B ± 0%      528B ± 0%   -4.35%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
Querify-4           18.0 ± 0%      16.0 ± 0%  -11.11%  (p=0.008 n=5+5)
QuerifySlice-4      23.0 ± 0%      21.0 ± 0%   -8.70%  (p=0.008 n=5+5)
```

What I've done is remove the allocation for adding items to pairedSlice, and define pairedSlice with specific length in both the cases like so:
```go
if len(params) == 1 && fmt.Sprintf("%v", reflect.TypeOf(params[0])) == "[]string" {
		pairedSlice := make([][]interface{}, 0, len(params[0].([]string)))
		if len(params[0].([]string))%2 != 0 {
			return "", errors.New("invalid query")
		}
		for i := 0; i < len(params[0].([]string)); i += 2 {
			pairedSlice = append(pairedSlice, []interface{}{params[0].([]string)[i], params[0].([]string)[i+1]})
		}
		for _, vals := range pairedSlice {
			qs.Add(fmt.Sprintf("%v", vals[0]), fmt.Sprintf("%v", vals[1]))
		}
	}
```
And similarly for the other case. I did try using the Dictionary function in one case (had to disable one of the false tests), but that didn't make too much of a difference except giving a positive delta in one case.