@moorereason 2021-03-14
Another hint:  refactor the entire function without using `pairedSlice`.  In both cases (len==1 and len>1), you already have access to the keys as strings.  Skip the `pairedSlice` step and use `qs.Add` directly in your loops.