@importhuman 2021-03-15
@moorereason Thanks, quite a notable difference now:
```
name            old time/op    new time/op    delta
Querify-4         3.25µs ± 3%    1.92µs ± 1%  -40.88%  (p=0.008 n=5+5)
QuerifySlice-4    3.96µs ± 2%    2.87µs ± 2%  -27.49%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
Querify-4           464B ± 0%      152B ± 0%  -67.24%  (p=0.008 n=5+5)
QuerifySlice-4      552B ± 0%      288B ± 0%  -47.83%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
Querify-4           18.0 ± 0%       9.0 ± 0%  -50.00%  (p=0.008 n=5+5)
QuerifySlice-4      23.0 ± 0%      17.0 ± 0%  -26.09%  (p=0.008 n=5+5)
```
Thinking about pointers to further improve performance, but not sure if/where I can use them.